+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This is the changelog for Tarmac.
 - Add `--log-level` argument to command
 - Add `--output-file` argument to command
 - Add `text` output format
+- Add `--version` argument to command
 
 ### Changed